Update to latest Scala-2.11.x and Scalatest
Note that there seems to be a slight difference in the ways scalatest-2.2.6 and
scalatest-3.0.3 treats expressions like

    foo shouldBe a [Bar]

for a value class `Bar`. With scalatest-3.0.3 checks of the above type might
fail, even when

    foo.isInstanceOf[Bar] shouldBe true

passes, because the implementation of the checks uses `ClassTag.runTimeClass`.

This commit adapts the tests accordingly.